COAR NotifyService IT failures: moving all /constriant calls to the new class, add log on the setup() to log deleted notifyService entities
